Fix compile errors.

+++ port/ingress_port.h
@@ -3,6 +3,6 @@
 
 class IngressPort {
   public:
-    virtual void** RxBurst(uint16_t burst_size) = 0;
+    virtual void** RxBurst(int burst_size) = 0;
 };
 #endif // _INGRESS_PORT_H_
