+++ port/null_egress_port.cc
@@ -1,5 +1,5 @@
 #include "null_egress_port.h"
 
-uint16_t NullEgressPort::TxBurst(void** packets, uint16_t burst_size) {
+int NullEgressPort::TxBurst(void** packets, int burst_size) {
   return 0;
 }
